+++ example/lib/textstyle_generator_trigger.dart
@@ -2,4 +2,4 @@ import 'package:flutter/material.dart';
 /// Imports the palette file that defines your default text color.
 import 'palette.dart';
 
-part 'generated/text_styles.g.dart';
+part 'generated_assets/text_styles.g.dart';